Move base routine into its own file
